eduplan/requirements.txt
fastapi==0.115.6
pyswip==0.3.2
uvicorn==0.32.1
pandas==2.2.1
python-dotenv==1.0.1
pymongo==4.3.3
requests==2.32.3